refactor scheduling config code path
