+++ apis/scheduling/config/v1beta2/doc.go
@@ -15,7 +15,7 @@ limitations under the License.
 */
 
 // +k8s:deepcopy-gen=package
-// +k8s:conversion-gen=github.com/koordinator-sh/koordinator/pkg/scheduler/apis/config
+// +k8s:conversion-gen=github.com/koordinator-sh/koordinator/apis/scheduling/config
 // +k8s:defaulter-gen=TypeMeta
 // +k8s:defaulter-gen-input=.
 // +groupName=kubescheduler.config.k8s.io
