+++ cmd/koord-scheduler/main.go
@@ -27,7 +27,7 @@ import (
 	"github.com/koordinator-sh/koordinator/pkg/scheduler/plugins/loadaware"
 
 	// Ensure scheme package is initialized.
-	_ "github.com/koordinator-sh/koordinator/pkg/scheduler/apis/config/scheme"
+	_ "github.com/koordinator-sh/koordinator/apis/scheduling/config/scheme"
 )
 
 func main() {
